+++ static/web/app.js
@@ -7,8 +7,8 @@
 
   var debug = window && window.location.toString().match(/.*\?debug/);
 
-  GpdvizController.$inject = ['$scope', '$window', '$timeout'];
-  function GpdvizController($scope, $window, $timeout) {
+  GpdvizController.$inject = ['$scope', '$window', '$timeout', '$http'];
+  function GpdvizController($scope, $window, $timeout, $http) {
     var vm = this;
     vm.debug = debug;
 
@@ -302,6 +302,35 @@
         vm.hoveredPoint = {};
         addSelectionPoint();
       });
+    });
+
+    map.on('click', function (e) {
+      if (!vm.ss || !vm.ss.clickListener) return;
+
+      var shiftKey = e.originalEvent.shiftKey;
+      var altKey   = e.originalEvent.altKey;
+      var metaKey  = e.originalEvent.metaKey;
+      // console.debug("clickListener=", vm.ss.clickListener, "MAP CLICK: e=", e
+      //   ,"latlng=", e.latlng
+      //   ,"shiftKey=", shiftKey
+      //   ,"altKey=",   altKey
+      //   ,"metaKey=",  metaKey
+      // );
+      if (shiftKey || altKey || metaKey) {
+        var params = {
+          lat:      e.latlng.lat,
+          lon:      e.latlng.lng,
+          shiftKey: shiftKey,
+          altKey:   altKey,
+          metaKey:  metaKey
+        };
+        $http({ method: "POST", url: vm.ss.clickListener, params: params}
+        ).then(function (response) {
+          // console.debug("url=", vm.ss.clickListener, "response=", response)
+        }, function (error) {
+          console.warn("url=", vm.ss.clickListener, "error=", error)
+        })
+      }
     });
 
     function createCharter(str) {